move Network/MemoryCacheDataSource to separate files
