+++ core/src/scene/sceneLoader.cpp
@@ -4,9 +4,11 @@
 #include "lights.h"
 #include "data/clientGeoJsonSource.h"
 #include "data/geoJsonSource.h"
+#include "data/memoryCacheDataSource.h"
 #include "data/mvtSource.h"
-#include "data/topoJsonSource.h"
+#include "data/networkDataSource.h"
 #include "data/rasterSource.h"
+#include "data/topoJsonSource.h"
 #include "gl/shaderProgram.h"
 #include "style/material.h"
 #include "style/polygonStyle.h"
